Added a fix to sessions.
Since sessions doesn't work with relative paths, I stoped using them.
But with this fix I can create a session that is cohesive with the
project that is being built. It means that I can create a .session file
and carry along other pc's. And the windows, tabs and buffer stack will
remain working, without dozens of warning signs.